--- Assets/Scenes/takifolder/field.cs
@@ -1,23 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class field : MonoBehaviour
-{
-    public int F;
-    //1岩場 2砂漠 3土地 4肥沃な土地 5草原 6森林 7海 0建造物
-    public int[] touchF = new int[4];
-    //隣のフィールドの状態
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
