+++ Assets/Scripts.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 524916d40d24b484fa37b6ea42bbe0fb
+guid: b0b826deaf2e003439d9c719edc5f383
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
