+++ Assets/Scripts/Creature.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d9a369d0839f9d440822b1330d6b134d
+guid: 2c1bc2657540dea44a0274bccb0a7169
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
