+++ UserSettings/EditorUserSettings.asset
@@ -11,6 +11,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-1:
       value: 22424703114646680e0b0227036c731118132e23293f68252320092a
       flags: 0
+    RecentlyUsedScenePath-2:
+      value: 22424703114646680e0b0227036c6b1505032c2b272168252320092a
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
